Fixed the issue with memory, used glushort to contain the indexes, since there are more than 256 indexes, glushort was not enough, so using glint, now a lot more # of indexes can be stored.

+++ PickAndTween/GLK2BufferObject.m
@@ -21,7 +21,7 @@
 -(void) uploadElementArray:(GLushort*)elements {
 
     glBindBuffer(self.glBufferType, self.glName);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(GLushort)*self.items, elements, GL_STATIC_DRAW);
+    glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(GLuint)*self.items, elements, GL_STATIC_DRAW);
 
 
 }
